Unify plotting tests to avoid repetition
Fix #448.
